+++ Changelog.md
@@ -4,6 +4,9 @@ Language Features:
 
 
 Compiler Features:
+ * Yul EVM Code Transform: Do not reuse stack slots that immediately become unreachable.
+ * Yul EVM Code Transform: Also pop unused argument slots for functions without return variables (under the same restrictions as for functions with return variables).
+ * Yul Optimizer: Move function arguments and return variables to memory with the experimental Stack Limit Evader (which is not enabled by default).
 
 
 Bugfixes: